requirements.txt
boto3==1.34.74
botocore==1.34.74
jmespath==1.0.1
numpy==1.26.4
pandas==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
